alertas de confirmacion y notificacion de errores, dialogo de detalles en el editor de notas y más opciones en el editor de notas

+++ src/app/pages/profile/profile.page.ts
@@ -8,6 +8,7 @@ import {AccountDeletionService} from "../../services/user/account-deletion.servi
 import {NotesService} from "../../services/notes/notes.service";
 import {ProfilePhotoService} from "../../services/user/profile-photo.service";
 import {AlertsService} from "../../services/alerts/alerts.service";
+import {ErrorsService} from "../../services/alerts/errors.service";
 
 @Component({
   selector: 'app-profile',
@@ -19,12 +20,17 @@ export class ProfilePage implements OnInit, OnDestroy {
   private userSubscription?: Subscription
   userNotesQuantity$?: Observable<number>
   userFavouriteNotesQuantity$?: Observable<number>
+  private readonly accountDeletionConfirmationMessage =
+    'Are you sure you want to delete your account and all of your notes? THIS ACTION CANNOT BE UNDONE.'
+  private readonly profilePhotoRemovalConfirmationMessage =
+    'Are you sure you want to remove your profile photo?'
 
   constructor(
     private userService: UserService,
     private authService: AuthService,
     private notesService: NotesService,
     private alertsService: AlertsService,
+    private errorsService: ErrorsService,
     private accountDeletionService: AccountDeletionService,
     private profilePhotoService: ProfilePhotoService,
     private router: Router
@@ -53,16 +59,50 @@ export class ProfilePage implements OnInit, OnDestroy {
       await this.authService.signOut()
       await this.router.navigate(['/home'])
     } catch (e: any) {
-      await this.alertsService.showErrorAlert(e.message)
+      await this.alertsService.showErrorAlert(this.errorsService.identifyError(e))
     }
   }
 
-  async onAccountDeletionConfirmationClosed(shouldDeleteAccount: boolean) {
-    if (shouldDeleteAccount)
-      await this.accountDeletionService.deleteUserAccount()
+  private async onAccountDeletionErrorAlertClosed() {
+    await this.authService.signOut()
+    await this.router.navigate(['/sign-in'])
   }
 
-  async onRemoveProfilePhotoButtonClicked() {
-    await this.profilePhotoService.deleteUserProfilePhoto()
+  private async onAccountDeletionConfirmationClosed(shouldDeleteAccount: boolean) {
+    try {
+      if (shouldDeleteAccount) {
+        await this.accountDeletionService.deleteUserAccount()
+        await this.router.navigate(['/home'])
+      }
+    } catch (e: any) {
+      await this.alertsService.showErrorAlert(
+        this.errorsService.identifyError(e),
+        this.onAccountDeletionErrorAlertClosed.bind(this)
+      )
+    }
   }
+
+  async onDeleteAccountButtonClicked() {
+    await this.alertsService.showDeleteConfirmationAlert(
+      this.accountDeletionConfirmationMessage,
+      this.onAccountDeletionConfirmationClosed.bind(this)
+    )
+  }
+
+  private async onProfilePhotoRemovalConfirmationClosed(shouldRemoveProfilePhoto: boolean) {
+    try {
+      if (shouldRemoveProfilePhoto)
+        await this.profilePhotoService.deleteUserProfilePhoto()
+    } catch (e: any) {
+      await this.alertsService.showErrorAlert(this.errorsService.identifyError(e))
+    }
+  }
+
+  async onProfilePhotoRemovalButtonClicked() {
+    await this.alertsService.showDeleteConfirmationAlert(
+      this.profilePhotoRemovalConfirmationMessage,
+      this.onProfilePhotoRemovalConfirmationClosed.bind(this)
+    )
+  }
+
 }
